backend/requirements.txt
<<<<<<< HEAD
<<<<<<< HEAD
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# AI/LLM
anthropic==0.7.7
openai==1.3.7

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
=======
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# OCR Dependencies
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
opencv-python==4.8.1.78
google-cloud-vision==3.5.0
boto3==1.34.0  # For AWS Textract

# Translation Dependencies
googletrans==4.0.0rc1
google-cloud-translate==3.12.1
anthropic==0.7.8
langdetect==1.0.9

# Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
>>>>>>> origin/claude/ocr-translation-modules-01Kv1eeHRaW9ea224g8V59NS

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
<<<<<<< HEAD
black==23.11.0
flake8==6.1.0
mypy==1.7.1
=======
# NEXUS Platform Backend Requirements
# Production-ready Python packages for FastAPI backend

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Async Task Queue
celery==5.3.6
redis==5.0.1

# AI - Anthropic Claude
anthropic==0.8.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
>>>>>>> origin/claude/build-campaign-manager-01KFpVYVNSuz2bfrpAMAZpwp
=======
black==23.12.0
ruff==0.1.8
mypy==1.7.1
>>>>>>> origin/claude/ocr-translation-modules-01Kv1eeHRaW9ea224g8V59NS
